# cli speed improvements - completion