--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -51,11 +51,12 @@ jobs:
 
       - name: Set Git author name and email address
         run: git config --global user.name "Bob The Builder" && git config --global user.email "bob@thebuilder.com"
-
-      - name: Run tests
-        run: | 
-          go get github.com/jbenet/go-context/io@v0.0.0-20150711004518-d14ea06fba99
+          
           # go get github.com/go-git/go-git/v5/plumbing/transport/ssh@v5.4.2
           # go get golang.org/x/net/idna@v0.0.0-20211015210444-4f30a5c0130f
           # go get github.com/AlecAivazis/survey/v2/terminal@v2.2.9
+      - name: Run tests
+        run: | 
+          go get github.com/jbenet/go-context/io@v0.0.0-20150711004518-d14ea06fba99
+          which ginkgo
           ginkgo -v ./...