--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -58,5 +58,4 @@ jobs:
       - name: Run tests
         run: | 
           go get github.com/jbenet/go-context/io@v0.0.0-20150711004518-d14ea06fba99
-          which ginkgo
-          ginkgo -v ./...
+          go test -v ./...