--- a/bob/aggregate.go
+++ b/bob/aggregate.go
@@ -57,6 +57,41 @@ func (b *B) PrintVersionCompatibility(bobfile *bobfile.Bobfile) {
 	}
 }
 
+// AggregateSparse reads Bobfile with the intent to gather task names.
+// The returned bobfile is not ready to be executed with a playbook.
+func (b *B) AggregateSparse() (aggregate *bobfile.Bobfile, err error) {
+	defer errz.Recover(&err)
+
+	bobfiles, err := b.find()
+	errz.Fatal(err)
+	bobs := []*bobfile.Bobfile{}
+
+	for _, bf := range bobfiles {
+		boblet, err := bobfile.BobfileReadPlain(filepath.Dir(bf))
+		errz.Fatal(err)
+
+		if boblet.Dir() == b.dir {
+			aggregate = boblet
+		}
+
+		bobs = append(bobs, boblet)
+	}
+
+	if aggregate == nil {
+		return nil, usererror.Wrap(ErrCouldNotFindTopLevelBobfile)
+	}
+
+	// set project names for all bobfiles and build tasks
+	aggregate, bobs = syncProjectName(aggregate, bobs)
+
+	aggregate.SetBobfiles(bobs)
+
+	// Merge tasks into one Bobfile
+	aggregate = b.addBuildTasksToAggregate(aggregate, bobs)
+
+	return aggregate, nil
+}
+
 // Aggregate determine and read Bobfiles recursively into memory
 // and returns a single Bobfile containing all tasks & runs.
 func (b *B) Aggregate() (aggregate *bobfile.Bobfile, err error) {
@@ -114,76 +149,15 @@ func (b *B) Aggregate() (aggregate *bobfile.Bobfile, err error) {
 	}
 
 	// set project names for all bobfiles and build tasks
-	for _, bobfile := range bobs {
-		bobfile.Project = aggregate.Project
-
-		for taskname, task := range bobfile.BTasks {
-			// Should be the name of the umbrella-bobfile.
-			task.SetProject(aggregate.Project)
-
-			// Overwrite value in build map
-			bobfile.BTasks[taskname] = task
-		}
-	}
+	syncProjectName(aggregate, bobs)
 
 	aggregate.SetBobfiles(bobs)
 
 	// Merge tasks into one Bobfile
-	for _, bobfile := range bobs {
-		// Skip the aggregate
-		if bobfile.Dir() == aggregate.Dir() {
-			continue
-		}
-
-		for taskname, task := range bobfile.BTasks {
-			dir := bobfile.Dir()
-
-			// Use a relative path as task prefix.
-			prefix := strings.TrimPrefix(dir, b.dir)
-			taskname := addTaskPrefix(prefix, taskname)
-
-			// Alter the taskname.
-			task.SetName(taskname)
-
-			// Rewrite dependent tasks to global scope.
-			dependsOn := []string{}
-			for _, dependentTask := range task.DependsOn {
-				dependsOn = append(dependsOn, addTaskPrefix(prefix, dependentTask))
-			}
-			task.DependsOn = dependsOn
-
-			aggregate.BTasks[taskname] = task
-		}
-	}
+	aggregate = b.addBuildTasksToAggregate(aggregate, bobs)
 
 	// Merge runs into one Bobfile
-	for _, bobfile := range bobs {
-		// Skip the aggregate
-		if bobfile.Dir() == aggregate.Dir() {
-			continue
-		}
-
-		for runname, run := range bobfile.RTasks {
-			dir := bobfile.Dir()
-
-			// Use a relative path as task prefix.
-			prefix := strings.TrimPrefix(dir, b.dir)
-
-			runname = addTaskPrefix(prefix, runname)
-
-			// Alter the runname.
-			run.SetName(runname)
-
-			// Rewrite dependents to global scope.
-			dependsOn := []string{}
-			for _, dependent := range run.DependsOn {
-				dependsOn = append(dependsOn, addTaskPrefix(prefix, dependent))
-			}
-			run.DependsOn = dependsOn
-
-			aggregate.RTasks[runname] = run
-		}
-	}
+	aggregate = b.addRunTasksToAggregate(aggregate, bobs)
 
 	// TODO: Gather missing tasks from remote  & Unpack?