--- a/bob/run_list.go
+++ b/bob/run_list.go
@@ -1,29 +1,15 @@
 package bob
 
 import (
-	"fmt"
 	"sort"
 
 	"github.com/benchkram/errz"
 )
 
-func (b *B) RunList() (err error) {
+func (b *B) GetRunTasks() (tasks []string, err error) {
 	defer errz.Recover(&err)
 
-	keys, err := b.GetRunList()
-	errz.Fatal(err)
-
-	for _, k := range keys {
-		fmt.Println(k)
-	}
-
-	return nil
-}
-
-func (b *B) GetRunList() (tasks []string, err error) {
-	defer errz.Recover(&err)
-
-	aggregate, err := b.Aggregate()
+	aggregate, err := b.AggregateSparse()
 	errz.Fatal(err)
 
 	keys := make([]string, 0, len(aggregate.RTasks))