--- a/cli/cmd_build.go
+++ b/cli/cmd_build.go
@@ -3,7 +3,6 @@ package cli
 import (
 	"context"
 	"errors"
-	"fmt"
 	"os"
 	"os/signal"
 	"strconv"
@@ -120,19 +119,4 @@ func getTasks() ([]string, error) {
 		return nil, err
 	}
 	return b.GetList()
-}
-
-// getTasksCmd cmd help to profile cli completion
-var getTasksCmd = &cobra.Command{
-	Use:   "gettasks",
-	Short: "gettasks",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		tasks, err := getTasks()
-		errz.Log(err)
-
-		for _, t := range tasks {
-			fmt.Println(t)
-		}
-	},
 }