This PR adds `AggregateSparse()` and uses it to get the tasknames faster. Intended to be used for cli completion. It doesn't call sanitize and removes most intitalisation overhead which is only required if doing a build.

Before:  1,29sec
After:     250ms

Hint: The initial `find()` function is responsible for about >95% (240ms) of execution time  (see flamegraph) only for searching child bobfiles. In case we implement **imports** the dimension is ~10ms for a medium size project. Though ~1ms would be ideal as much bigger projects exist.. no master plan (yet) how to get there.

![image](https://user-images.githubusercontent.com/3776893/166069629-5f7af98a-a8d8-466a-8419-13eb79d97d42.png)